% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiators.R
\name{case_control_sampling}
\alias{case_control_sampling}
\title{Case-control sampling from extended data}
\usage{
case_control_sampling(data_dir, n_control, numCores = 1)
}
\arguments{
\item{data_dir}{Directory to read 'sw_data.csv' from}

\item{n_control}{Number of controls for each case to sample}

\item{numCores}{Number of cores used by mclapply for sampling from each trial}
}
\description{
Case-control sampling from extended data
}
