% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiators.R
\name{case_control_sampling_trials}
\alias{case_control_sampling_trials}
\title{Case-control sampling from extended data in separate trial csvs}
\usage{
case_control_sampling_trials(
  data_dir,
  n_control,
  numCores,
  name_prefix = "cc_sample_",
  infile_pattern = "trial_"
)
}
\arguments{
\item{data_dir}{Directory containing trial-level extended data}

\item{n_control}{Number of controls to sample per case}

\item{numCores}{Number of cores for parallel processing}
}
\value{

}
\description{
Case-control sampling from extended data in separate trial csvs
}
