% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiators.R
\name{case_control_sampling}
\alias{case_control_sampling}
\title{Case-control sampling from extended data}
\usage{
case_control_sampling(data_dir, n_control, numCores)
}
\value{

}
\description{
Case-control sampling from extended data
}
